Moved ProductStoreSelector and ProductIdentifier inline in Settings page Removed Unused Files
